+++ app/components/views/TicketsPage/VSPTicketsStatusTab/MyTicketsTab.module.css
